requirements.txt
djangorestframework
django-filter
django==1.7
markdown
pil
jwt